refactor: Parser의 validation 메서드 Validator로 분리
